refactor(comments): refactor focus handling in the CommentInput
This will fix some bugs we have been seeing related to setting focus in the comment input.
Use the PTE Editable selection prop to control selecting end of the content.
Call focus on the editor without side-effects.
