+++ packages/sanity/src/desk/comments/src/components/pte/comment-input/CommentInputProvider.tsx
@@ -16,7 +16,6 @@ export interface CommentInputContextValue {
   expandOnFocus?: boolean
   focused: boolean
   focusEditor: () => void
-  focusEditorEndOfContent: () => void
   focusOnMount?: boolean
   hasChanges: boolean
   insertAtChar: () => void
@@ -63,32 +62,6 @@ export function CommentInputProvider(props: CommentInputProviderProps) {
   const hasChanges = useCommentHasChanged(value)
 
   const focusEditor = useCallback(() => PortableTextEditor.focus(editor), [editor])
-
-  const focusEditorEndOfContent = useCallback(() => {
-    const _value = PortableTextEditor.getValue(editor)
-    const lastBlock = (_value || []).slice(-1)[0]
-    if (!lastBlock) {
-      PortableTextEditor.focus(editor)
-      return
-    }
-    const lastChild = isPortableTextTextBlock(lastBlock)
-      ? lastBlock.children.slice(-1)[0]
-      : undefined
-    if (!lastChild) {
-      PortableTextEditor.focus(editor)
-      return
-    }
-    const point = {
-      path: [{_key: lastBlock._key}, 'children', {_key: lastChild._key}],
-      offset: isPortableTextSpan(lastChild) ? lastChild.text.length : 0,
-    }
-    const newSelection = {
-      focus: point,
-      anchor: point,
-    }
-    PortableTextEditor.select(editor, newSelection)
-    PortableTextEditor.focus(editor)
-  }, [editor])
 
   const closeMentions = useCallback(() => {
     setMentionsMenuOpen(false)
@@ -236,7 +209,6 @@ export function CommentInputProvider(props: CommentInputProviderProps) {
         expandOnFocus,
         focused,
         focusEditor,
-        focusEditorEndOfContent,
         focusOnMount,
         hasChanges,
         insertAtChar,
@@ -255,7 +227,6 @@ export function CommentInputProvider(props: CommentInputProviderProps) {
       expandOnFocus,
       focused,
       focusEditor,
-      focusEditorEndOfContent,
       focusOnMount,
       hasChanges,
       insertAtChar,
